Complete SpannerPartitionReader*
This change implements SpannerPartitionReader which takes
in an InputPartition which would have been generated after
running a partitionQuery that then serializes the returned
Partition over the network to the executor which when creating
SpannerPartitionReader acquires a transaction then retrieves
values from the produced ResultSet.
Later on, we shall need to investigate if the transaction
might be held for longer than 10 seconds and might be idle
for that long.